+++ CHANGES.md
@@ -1,8 +1,9 @@
 **Added:**
-* none
+* control class update group
 
 **Changed:**
 * none
 
 **Fixed:**
-* none
+* canvas controls: step labels weren't being displayed properly on small screen sizes (... was inside label text)
+* project init window: fixed "create new project" button size not being big enough for small screen sizes